Support/pacfix

+++ lib/cpr_runtime.c
@@ -170,7 +170,7 @@ void init_tables() {
 }
 
 int __cpr_choice(char* lid, char* typestr,
-                     int* rvals, char** rvals_ids, int rvals_size,
+                     long long* rvals, char** rvals_ids, int rvals_size,
                      int** lvals, char** lvals_ids, int lvals_size){
     if (!choice_instances)
         init_tables();
